+++ src/layout/sidebar/LogoContainer.ts
@@ -1,6 +1,0 @@
-import styled from "../../styled-components";
-
-export const LogoContainer = styled.div`
-    padding-left: 72px;
-    padding-right: 56px;
-`;