+++ src/layout/sidebar/SidebarPageTitle.ts
@@ -1,8 +1,0 @@
-import styled from "../../styled-components";
-
-export const SidebarPageTitle = styled.div`
-    font-size: 14px;
-    display: inline-block;
-    align-self: center;
-    color: ${({theme}) => theme.colors.sidebarPageTitle};
-`;